% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.daily_aggregate}
\alias{.daily_aggregate}
\title{Aggregate data.frame Object}
\usage{
.daily_aggregate(df, FUN = function(x) {
    mean(x, na.rm = TRUE)
})
}
\arguments{
\item{df}{\emph{data.frame} object.}

\item{FUN}{The function to be applied to each vector of numeric data.}
}
\value{
Returns an aggregated \emph{data.frame} object.
}
\description{
Aggregate (\emph{data.frame}) object along its
datetime axis. Temporal aggregation involves splitting a \emph{data.frame} object into
separate bins along its datetime axis. \code{FUN} is mapped to the \emph{data.frame}
numeric variables in each bin.
}
\details{
\code{FUN} must operate on univariate numeric vectors and return a
scalar value. Besides the data variable, no additional arguments will be
provided to this function. This means that functions like \code{mean} and
\code{max} will need to be wrapped in a function that specifies
\code{na.rm = TRUE}. See the examples below.
}
\keyword{internal}
