% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_airsensor_object}
\alias{example_airsensor_object}
\title{Example AirSensor Timeseries dataset}
\format{
An S3 object composed of "meta" and "data" data.
}
\usage{
example_airsensor_object
}
\description{
The \code{example_airsensor_object} dataset provides a quickly
loadable version of an \pkg{AirSensor} package \emph{airsensor} object for
code examples.
This dataset was generated with AirSensor version 0.9.17 on 2020-09-16 by
running:

\preformatted{
library(AirSensor)

setArchiveBaseUrl("http://data.mazamascience.com/PurpleAir/v1")

example_airsensor_object <-
  AirSensor::sensor_load(startdate = 20200901, enddate = 20200915)

save(example_airsensor_object, file = "data/example_airsensor_object.rda")
}
}
\keyword{datasets}
