% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pwfslsmoke_object}
\alias{example_pwfslsmoke_object}
\title{Example PWFSLSmoke Timeseries dataset}
\format{
An S3 object composed of "meta" and "data" data.
}
\usage{
example_pwfslsmoke_object
}
\description{
The \code{example_pwfslsmoke_object} dataset provides a quickly
loadable version of an \pkg{PWFSLSmoke} package \emph{ws_monitor} object for
code examples.
This dataset was generated with PWFSLSmoke version 1.2.112 on 2020-08-14 by
running:

\preformatted{
library(PWFSLSmoke)

}
}
\keyword{datasets}
