% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor.R
\docType{data}
\name{sensor}
\alias{sensor}
\title{Example sensor dataset}
\format{
A list of "meta" data.frame (1 row 4 variables) and a "data"
data.frame(566 rows 2 variables).
\describe{
\item{meta$id}{A unique identifier.}
\item{meta$label}{A human label of the identifier.}
\item{meta$longitide}{A Longitude.}
\item{meta$latitude}{A Latitude.}
\item{data$date}{A regular POSIXct timeseries}
\item{data$example_id}{A id matching the corresponding meta$id}
}
}
\usage{
sensor
}
\description{
A dataset containing a list of two dataframes, "meta" and "data". "meta"
contains an index, label, longitude, and latitude, while "data" contains
regular univariate temporal data.
}
\keyword{datasets}
