% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesBarChart.R
\name{timeseriesBarChart}
\alias{timeseriesBarChart}
\title{<Add Title>}
\usage{
timeseriesBarChart(
  meta,
  data,
  shared = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
