% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiotemp.R
\docType{package}
\name{tiotemp}
\alias{tiotemp}
\title{Spa(tiotemp)oral Visualization}
\description{
The tiotemp package provides a set of useful timeseries and spatial
visualizations designed to be viewed in the RStudio Viewer pane or
or incorporated into Shiny apps..
}
