% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesBarChart.R
\name{timeseriesBarChart}
\alias{timeseriesBarChart}
\title{<Add Title>}
\usage{
timeseriesBarChart(data, meta, index = "monitorID", label = "label", ...)
}
\description{
<Add Description>
}
\examples{
library(AirSensor)
sensor <- example_sensor
timeseriesBarChart(meta = sensor$meta, data = sensor$data)
}
