% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesBarChart.R
\name{timeseriesBarChart}
\alias{timeseriesBarChart}
\title{<Add Title>}
\usage{
timeseriesBarChart(
  meta,
  data,
  title = NULL,
  subtitle = NULL,
  ylab = "PM₂.₅ (μg/m³)",
  shared = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
\examples{
library(AirSensor)
sensor <- example_sensor
timeseriesBarChart(meta = sensor$meta, data = sensor$data)
}
