% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesBarChart.R
\name{timeseriesBarChart}
\alias{timeseriesBarChart}
\title{Timeseries Bar Chart}
\usage{
timeseriesBarChart(data, meta, index = "monitorID", label = "label", ...)
}
\arguments{
\item{data}{A data.frame that contains the hourly-resolution point location
timeseries data. See Details.}

\item{meta}{A data.frame that contains the point location metadata. See Details.}

\item{index}{A string to index point location metadata and temporal data by.}

\item{label}{A string to index point location metadata label by.}

\item{...}{Additional arguments. See details.}
}
\description{
A timeseries barchart.
}
\examples{
library(tiotemp)
timeseriesBarChart(data = sensor$data, meta = sensor$meta)
}
