% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesCalendar.R
\name{timeseriesCalendar}
\alias{timeseriesCalendar}
\title{Timeseries Calendar Heatmap}
\usage{
timeseriesCalendar(
  data,
  meta,
  index = "monitorID",
  label = "label",
  full = TRUE,
  ...
)
}
\arguments{
\item{data}{A data.frame that contains the hourly-resolution point location
timeseries data. See Details.}

\item{meta}{A data.frame that contains the point location metadata. See Details.}

\item{index}{A string to index point location metadata and temporal data by.}

\item{label}{A string to index point location metadata label by.}

\item{full}{Fill and display full calendar year.}

\item{...}{Additional arguments. See details.}
}
\description{

}
\examples{
library(tiotemp)

sensor <- example_airsensor_object

timeseriesCalendar(data = sensor$data[,c(1,2)], meta = sensor$meta[1,])

}
