% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseriesCalendar.R
\name{timeseriesCalendar}
\alias{timeseriesCalendar}
\title{<Add Title>}
\usage{
timeseriesCalendar(data, meta, index = "monitorID", label = "label", ...)
}
\description{
<Add Description>
}
\examples{
library(AirSensor)
sensor <- example_sensor
timeseriesCalendar(meta = sensor$meta, data = sensor$data)
}
